Add solution for dot product
Working from the sketch in dot.c, this commit adds a solution to a dot
product of two arrays written in RISCV assembly. Note the use of calling
conventions: the values in the callee-saved registers are pushed onto
the stack during the function prologue, then the registers are used to
store local variables within the function. The callee-saved registers
are then popped back off the stack during the epilogue, after the return
value has been loaded into the return register.